introduces post.js and refactors reconnectcounter
